+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## feature/elections-and-map-management
 ### Fix
  - Add a visible json for all departements readable by API
+### Add
+ - Mongoose models to describe elections
 
 ## branch refacto/test_fix_and_improve_user_features_and_middlewares
 ### Fix